Fix VisualizationResult violations
