+++ frontend/src/metabase/query_builder/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from "./use-breakout-query-handlers";
 export * from "./use-default-query-aggregation";
+export * from "./use-visualization-result-qb-props";
 export * from "./types";
